for 安全检查图标

+++ Syberos/syberosclient/src/main/java/com/syberos/shuili/fragment/thematic/detail/ThematicDetailSinsFragment.java
@@ -183,14 +183,14 @@ public class ThematicDetailSinsFragment extends BaseLazyFragment {
 
 
             List<PieEntry> listHiddenRate = new ArrayList<>();
-            listHiddenRate.add(new PieEntry(sinsEntry.getData().getSINSHIDDRECTQUA(), "已检查 " + sinsEntry.getData().getSINSHIDDRECTQUA() + ""));
-            listHiddenRate.add(new PieEntry(sinsEntry.getData().getSINSHIDDQUA() - sinsEntry.getData().getSINSHIDDRECTQUA(), "未检查 " + (sinsEntry.getData().getSINSHIDDQUA() - sinsEntry.getData().getSINSHIDDRECTQUA()) + ""));
+            listHiddenRate.add(new PieEntry(Float.valueOf(sinsEntry.getData().getSINSHIDDRECTQUA()), "已检查 " + sinsEntry.getData().getSINSHIDDRECTQUA() + ""));
+            listHiddenRate.add(new PieEntry(Float.valueOf(sinsEntry.getData().getSINSHIDDQUA()) - Float.valueOf(sinsEntry.getData().getSINSHIDDRECTQUA()), "未检查 " + (Integer.valueOf(sinsEntry.getData().getSINSHIDDQUA())- Integer.valueOf(sinsEntry.getData().getSINSHIDDRECTQUA())) + ""));
 
             MPChartUtil.getInstance().initPieCharHiddenRate(mContext, pieCharSinsRate, listHiddenRate, true);
 
             List<ProjectEntry> list = new ArrayList<>();
             for (SinsEntry.DataBean.SUBSINSDATABean bean : sinsEntry.getData().getSUBSINSDATA()) {
-                list.add(new ProjectEntry(bean.getOBJGUID(), bean.getOBJNAME(), bean.getSINSHIDDQUA()));
+                list.add(new ProjectEntry(bean.getOBJGUID(), bean.getOBJNAME(), Integer.valueOf(bean.getSINSHIDDQUA())));
             }
             RecyclerAdapterGeneral adapter = new RecyclerAdapterGeneral(list,"个");
             recyclerView.setLayoutManager(new LinearLayoutManager(mContext));
